refactor(sdk): Space module
