+++ sdk/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod aggregations;
+pub mod aggregation;
 pub mod error;
 pub mod graph_uri;
 pub mod ids;
