+++ sdk/src/models/space/mod.rs
@@ -1,0 +1,17 @@
+pub mod find_one_query;
+pub mod find_many_query;
+pub mod parent_spaces_query;
+pub mod space;
+pub mod space_editors_query;
+pub mod space_members_query;
+pub mod space_types_query;
+pub mod subspaces_query;
+
+pub use find_one_query::FindOneQuery;
+pub use find_many_query::FindManyQuery;
+pub use parent_spaces_query::ParentSpacesQuery;
+pub use space::{ParentSpace, Space, SpaceBuilder, SpaceGovernanceType};
+pub use space_editors_query::SpaceEditorsQuery;
+pub use space_members_query::SpaceMembersQuery;
+pub use space_types_query::SpaceTypesQuery;
+pub use subspaces_query::SubspacesQuery;
